metrics: datapoint macro: Allow trailing comma (#30317)
Rust grammar allows trailing commas in most places where a list of
elements are accepted.  It simplifies cases when the list is generated
by a macro, allowing the macro to avoid special cases for a one element
list vs longer lists.

As such, it is a common practice to allow trailing commas in macros as
well.