sdk: Extract `solana-sysvar-id` crate (#3309)
* sdk: Extract `solana-sysvar` crate

#### Problem

The sysvars in `solana-program` are tightly coupled with types that
exist in `solana-program`. For example, all of the special sysvar
getters like `Rent::get()` are implemented through a macro that falls
back to using `program_stubs`.

Because of this tight coupling, it's difficult to pull out bits for the
sysvars.

#### Summary of changes

After numerous attempts, I've decided to keep it simple and only extract
`SysvarId`, its helper macros, and `get_sysvar`.

To go along with that, all of the separated sysvar crates now implement
the sysvar ids themselves under a new `sysvar` feature. This new feature
might be overkill, so let me know if we should just include the sysvar
ids by default. I went with a feature to include an implementation using
the `sol_get_sysvar` syscall in the future.

It was really messy to include the `Sysvar` trait from `solana-program`
because it falls back to using `bincode`, which we know performs poorly
for on-chain programs. So the future idea is to create a new `Sysvar`
trait in `solana-sysvar` which will require fewer bits to deserialize
sysvars.

Let me know what you think about this PR and the future idea! Note that
I'll need to rebase this on top of #3249 and #3272 when they land.

* Add solana-define-syscall for solana builds

* Rename solana-sysvar -> solana-sysvar-id

* Move back `get_sysvar` to solana-program

* Update lockfile for v2.2