sdk: Rename ed25519-instructions -> ed25519-program (#3632)
* sdk: Rename ed25519-instructions -> ed25519-program

#### Problem

The precompiles behave more like programs than instructions, as was
decided with the new secp256r1 precompile in #3152, but the ed25519
precompile was split as `ed25519-instructions`.

#### Summary of changes

Rename the crate from `solana-ed25519-instructions` to
`solana-ed25519-program`, which is a previous crate that used to be used
until https://github.com/solana-labs/solana/pull/19930.

* Run cargo fmt