tls-utils: Make publishable, remove sdk dep (#3909)
#### Problem

The new tls-utils crate is a great extraction of re-used logic, but it
wasn't marked publishable, even though it will need to be published due
to transitive dependencies from client crates.

Also, it has an unnecessary dependency on solana-sdk.

#### Summary of changes

Remove the solana-sdk dependency and use solana-keypair directly. And
make the crate publishable.