sdk: Extract instructions-sysvar (#4040)
* sdk: Extract instructions-sysvar

#### Problem

The instructions sysvar currently lives in the `solana-sysvar` crate,
but it doesn't require any of the types from the sysvar crate, making it
a great candidate for extraction.

#### Summary of changes

Since `solana-sysvar` hasn't been officially published yet, we can make
a breaking change to it, and totally extract the instructions sysvar
from it, then re-export the functionality from `solana-program`
directly.

* Fix secp256k1 doctests

* Sort deps