deps: Relax spl dependencies (#5704)
* deps: Relax spl dependencies

#### Problem

The sdk dependencies are now relaxed for non-binary crates in Agave, but
the spl crates are still pinned, which can lead to downstream build
issues like
[this CI run](https://github.com/anza-xyz/agave/actions/runs/14316385404/job/40123415833?pr=5629).

#### Summary of changes

Relax the SPL dependencies in the whole agave repo. Note that I didn't
pin binary crates with direct non-dev dependencies on SPL. The list
includes:

* bench-tps (for instruction-padding)
* CLI (for spl-memo)
* faucet (for spl-memo)
* tokens (for spl-token)

Since these are less mission-critical binaries, it seemed reasonable to
avoid the pinning, but let me know if you feel differently.

* Re-enable downstream jobs

* Pin spl versions in bin deps