Use TaskTracket in streamer for graceful exit and avoiding flaky tests (#8066)
When streamer service is launched, it spawns task and returns handle for this task. This task internally spawns tasks to handle connections without tracking them.
Which means that when the root task has stopped we have no idea if it's child tasks have completed or not.
This leads to flaky tests. For example, in nonblocking::quic::test::test_quic_server_multiple_streams we stop the streamer and after that check stats.total_connections.load(Ordering::Relaxed). This counter counts how many task are handling at giving moment connections. It might happen that these tasks haven't finished before we do this check, so all the checks below task_handle.await are potentially flaky.